+++ examples/tutorial/basic-2/programs/basic-2/src/lib.rs
@@ -1,5 +1,3 @@
-#![feature(proc_macro_hygiene)]
-
 use anchor_lang::prelude::*;
 
 // Define the program's instruction handlers.