+++ examples/tutorial/basic-4/programs/basic-4/src/lib.rs
@@ -1,5 +1,3 @@
-#![feature(proc_macro_hygiene)]
-
 // #region code
 use anchor_lang::prelude::*;