+++ examples/typescript/programs/typescript/src/lib.rs
@@ -1,8 +1,6 @@
 //! The typescript example serves to show how one would setup an Anchor
 //! workspace with TypeScript tests and migrations.
 
-#![feature(proc_macro_hygiene)]
-
 use anchor_lang::prelude::*;
 
 #[program]