+++ CHANGELOG.md
@@ -14,6 +14,7 @@ incremented for features.
 ## Features
 
 * cli: Specify test files to run ([#118](https://github.com/project-serum/anchor/pull/118)).
+* lang: Allow overriding the `#[state]` account's size ([#121](https://github.com/project-serum/anchor/pull/121)).
 
 ## [0.3.0] - 2021-03-12